requirements.txt
# Updated requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
requests==2.31.0  # Added for M-Pesa API calls
